Added more constants, also minor refactoring.
Complex: added constant i;
Constants: moved from Number module to Field; also added gamma, phi.
