+++ src/main/scala/com/phasmidsoftware/number/core/Number.scala
@@ -1024,56 +1024,6 @@ object Number {
   }
 }
 
-object Constants {
-  val sG = "6.67430(15)E-11" // m ^ 3 kg ^ -1 s ^ -2
-  val sAlpha = "0.0072973525693(11)" // (dimensionless)
-  val sAvagadro = "6.0221407600E23" // mole ^ -1
-  val sBoltzmann = "1.38064900E-23" // J K ^ -1
-  val sPlanck = "6.6260701500E-34" // J Hz ^ -1
-  val sC = "299792458" // m sec ^ -1
-  val sMu = "1836.15267343(11)" // (dimensionless)
-
-  val one: Number = Number.one
-  val zero: Number = Number.zero
-  val pi: Number = Number.pi
-  val e: Number = Number.e
-
-  /**
-    * https://en.wikipedia.org/wiki/Gravitational_constant
-    */
-  lazy val G: Number = Number(sG)
-
-  /**
-    * https://en.wikipedia.org/wiki/Fine-structure_constant
-    */
-  lazy val alpha: Number = Number(sAlpha)
-
-  /**
-    * https://en.wikipedia.org/wiki/Avogadro_constant
-    */
-  lazy val avagadro: Number = Number(sAvagadro)
-
-  /**
-    * https://en.wikipedia.org/wiki/Boltzmann_constant
-    */
-  lazy val boltzmann: Number = Number(sBoltzmann)
-
-  /**
-    * https://en.wikipedia.org/wiki/Planck_constant
-    */
-  lazy val planck: Number = Number(sPlanck)
-
-  /**
-    * https://en.wikipedia.org/wiki/Speed_of_light
-    */
-  lazy val c: Number = Number(sC)
-
-  /**
-    * https://en.wikipedia.org/wiki/Proton-to-electron_mass_ratio
-    */
-  lazy val mu: Number = Number(sMu)
-}
-
 case class NumberException(str: String) extends Exception(str)
 
 case class NumberExceptionWithCause(str: String, e: Throwable) extends Exception(str, e)
